% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_db_remea_scores.R
\name{combine_remea_scores}
\alias{combine_remea_scores}
\title{Combine ReMEA score}
\usage{
combine_remea_scores(remea_results)
}
\arguments{
\item{remea_results}{Output from response_marker_enrichment_analysis function.}
}
\value{
`data.table` of combined ReMEA scores across signature DBs.
}
\description{
Function to average ReMEA scores for perturbagens across different datasets.
}
\examples{
combine_remea_score(remea_results = remea_results_dt)
}
