% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_response_marker_enrichment_analysis.R
\name{CTAM_signature_enrichment}
\alias{CTAM_signature_enrichment}
\title{Signature enrichment CTAMS}
\usage{
CTAM_signature_enrichment(protein_data, signatures, signature.type)
}
\arguments{
\item{protein_data}{`data.frame` Proteomics differential testing results. Protein name
must be included as protein_acc. Fold change data must be in column 2.}

\item{signatures}{`data.table` Signature DB with response protein signatures.
This can be found in the ReMEA NAMESPACE. e.g. ReMEA::list_of_signature_lists}

\item{signature.type}{`string` Naming variable for signatures. Main purpose if for
response_marker_enrichment_analysis, where this indicates signature DB ids.}
}
\value{
`data.table` with enrichment results
}
\description{
CTAM marker enrichment function for ReMEA pipeline.
}
\examples{
signature_enrichemnt(protein_data = protein_data, signatures  = signature_dt, signature.type = "signature_id")
}
