% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remea_analysis.R
\name{get_ReMEA_scores}
\alias{get_ReMEA_scores}
\title{Get averaged ReMEA scores}
\usage{
get_ReMEA_scores(
  protein_data,
  tumour_type = c("pan"),
  marker_type = "DRUG",
  signature_version = "CellLines",
  protein_id_col = NULL,
  analysis_col = NULL,
  return_individual_scores = FALSE
)
}
\arguments{
\item{protein_data}{A data.table containing proteomics differential-testing
results. A data.frame is also accepted and will be converted to a
data.table. Proteins must be annotated with their UniProt Entry Names.
The protein identifier column can be specified using protein_id_col; if
omitted, the first column is used. The numeric column to analyse, such as
fold change, can be specified using analysis_col; if omitted, the second
column is used.}

\item{tumour_type}{`string` Whether to use markers from correlation analysis of solid tumours
or non-solid tumours only. If multiple passed, these will be analysed in looped format.}

\item{marker_type}{`string` Perturbagen type for analysis. These are CRISPR, RNAi or DRUG.
In case of selecting both RNAi and CRISPR, these may be combined using the `combine_remea_score` function.}

\item{signature_version}{`string` Signature version to use.}

\item{protein_id_col}{Column for protein identities. To use package signatures,
these must refer to UniProt entry name of proteins. If NULL, will default to first column.}

\item{analysis_col}{Data column to analysis. If NULL, will default to second column.}

\item{return_individual_scores}{`logical` whether to return results for individual db scores.}
}
\value{
`data.table` of combined ReMEA scores if return_individual_scores == FALSE,
else returns list of both combined and individual scores.
}
\description{
Get averaged ReMEA scores
}
