% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_mean.R
\name{gm_mean}
\alias{gm_mean}
\title{Geometric mean}
\usage{
gm_mean(x, na.rm = TRUE)
}
\arguments{
\item{x}{numeric vector.}

\item{na.rm}{`logical` whether or not to exclude na values.}
}
\value{
`float` Geometric mean of x.
}
\description{
Geometric mean
}
