% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_db_remea_scores.R
\name{pcom_by_stouffer}
\alias{pcom_by_stouffer}
\title{pcom by stouffer}
\usage{
pcom_by_stouffer(p)
}
\arguments{
\item{p}{pvalue}
}
\value{
combined p values.
}
\description{
pcom by stouffer
}
