% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_remea.R
\name{plot_crispr_vs_rnai}
\alias{plot_crispr_vs_rnai}
\title{Plot CRISPR versus RNAi ReMEA scores}
\usage{
plot_crispr_vs_rnai(
  remea_scores = NULL,
  combined_RNAi = NULL,
  combined_CRISPR = NULL,
  drug_target_genes = FALSE
)
}
\arguments{
\item{remea_scores}{`list` output from `complete_ReMEA_analysis` function.}

\item{combined_RNAi}{`data.table` of combined RNAi ReMEA scores.}

\item{combined_CRISPR}{`data.table` of combined CRISPR ReMEA scores.}

\item{drug_target_genes}{`logical` whether to filter plot to just include genes of drug targets.}
}
\value{
ggplot object
}
\description{
Plot CRISPR versus RNAi ReMEA scores
}
