% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_remea.R
\name{plot_drug_vs_gene}
\alias{plot_drug_vs_gene}
\title{Plot drug ReMEA scores versus gene target perturbation ReMEA scores}
\usage{
plot_drug_vs_gene(
  remea_scores = NULL,
  combined_drug = NULL,
  combined_gene = NULL,
  gene_perturbation_type = NULL
)
}
\arguments{
\item{remea_scores}{`list` output from `complete_ReMEA_analysis` function.}

\item{combined_drug}{`data.table` of combined RNAi ReMEA scores.}

\item{combined_gene}{`data.table` of combined CRISPR ReMEA scores.}

\item{gene_perturbation_type}{`string` indicating whether "RNAi", "CRISPR" or "RNAi_CRISPR" scores should be extracted from `remea_scores` list.}
}
\value{
ggplot object
}
\description{
Plot drug ReMEA scores versus gene target perturbation ReMEA scores
}
