% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_remea.R
\name{plot_individual_db_ReMEA_scores}
\alias{plot_individual_db_ReMEA_scores}
\title{Plot individual database max delta score}
\usage{
plot_individual_db_ReMEA_scores(remea_db_scores, drug_target_genes = FALSE)
}
\arguments{
\item{remea_db_scores}{`data.table` of individual db scores. Output of `response_marker_enrichment_analysis` function.}

\item{drug_target_gens}{`logical` Whether to plot only drug target genes.}
}
\value{
ggplot object
}
\description{
Plot individual database max delta score
}
