% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_remea.R
\name{plot_ReMEA_scores}
\alias{plot_ReMEA_scores}
\title{Plot combined ReMEA score}
\usage{
plot_ReMEA_scores(combined_remea, drug_target_genes = FALSE)
}
\arguments{
\item{combined_remea}{`data.table` of combined ReMEA scores. Output of `combine_remea_scores` function.}

\item{drug_target_genes}{`logical` Whether to plot only drug target genes.}
}
\value{
ggplot object
}
\description{
Plot combined ReMEA score
}
