% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_remea.R
\name{plot_ReMEA_volcano}
\alias{plot_ReMEA_volcano}
\title{Plot ReMEA score volcano}
\usage{
plot_ReMEA_volcano(combined_remea, qval_cut = 0.05, drug_target_genes = FALSE)
}
\arguments{
\item{combined_remea}{`data.table` of combined ReMEA scores. Output of `combine_remea_scores` function.}

\item{qval_cut}{`float` labelling cut off for volcano plot.}

\item{drug_target_genes}{`logical` whether to just plot all genes or just drug target genes.}
}
\value{
ggplot object
}
\description{
Plot ReMEA score volcano
}
