% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReMEA-package.R
\docType{package}
\name{ReMEA-package}
\alias{ReMEA}
\alias{ReMEA-package}
\title{ReMEA: Response Marker Enrichment Analysis (Title Case)}
\description{
Created on 15/08/2023, this iteration revised major functions, added roxygen comments and included pictilisib correlated genes.
}
\author{
\strong{Maintainer}: Luke Higgins \email{l.higgins@qmul.ac.uk}

Authors:
\itemize{
  \item Luke Higgins \email{l.higgins@qmul.ac.uk}
  \item Pedro Rodriguez Cutillas \email{p.cutillas@qmul.ac.uk}
}

}
\keyword{internal}
