% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_remea.R
\name{remea_dotplots}
\alias{remea_dotplots}
\title{ReMEA dotplot panel}
\usage{
remea_dotplots(
  combined_remea,
  remea_db_scores,
  drug_target_genes = FALSE,
  topn = 25
)
}
\arguments{
\item{combined_remea}{`data.table` of combined ReMEA scores. Output of `combine_remea_scores` function.}

\item{drug_target_genes}{`data.table` of combined ReMEA scores. Output of `combine_remea_scores` function.}

\item{topn}{`integer` number of top perturbations to plot.}

\item{individual_db_scores}{`data.table` of individual db scores. Output of `response_marker_enrichment_analysis` function.}
}
\description{
ReMEA dotplot panel
}
