% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_remea.R
\name{remea_plots}
\alias{remea_plots}
\title{ReMEA plot panel}
\usage{
remea_plots(individual_db_scores, combined_remea, drug_target_genes = FALSE)
}
\arguments{
\item{individual_db_scores}{`data.table` of individual db scores. Output of `response_marker_enrichment_analysis` function.}

\item{combined_remea}{`data.table` of combined ReMEA scores. Output of `combine_remea_scores` function.}

\item{drug_target_genes}{`data.table` of combined ReMEA scores. Output of `combine_remea_scores` function.}
}
\description{
ReMEA plot panel
}
