% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_response_marker_enrichment_analysis.R
\name{response_marker_enrichment_analysis_APROCRITA}
\alias{response_marker_enrichment_analysis_APROCRITA}
\title{Response marker enrichment analysis APOCRITA VERSION}
\usage{
response_marker_enrichment_analysis_APROCRITA(
  protein_data,
  marker_type = c("DRUG", "RNAi", "CRISPR"),
  tumour_type = c("haem", "solid"),
  signature_version = "CellLines",
  dataset_selection = FALSE,
  proteomics_datasets = NULL,
  perturbation_datasets = NULL
)
}
\arguments{
\item{protein_data}{`data.frame` Proteomics differential testing results. Protein name
must be included as protein_acc. Fold change data must be in column 2.}

\item{marker_type}{`string` Perturbagen type for analysis. These are CRISPR, RNAi or DRUG.
Multiple markers may be passed. In case of selecting both RNAi and CRISPR, these
may be combined using the `combine_remea_score` function}

\item{tumour_type}{`string` Whether to use markers from correlation analysis of solid tumours
or non-solid tumours only. If multiple passed, these will be analysed in looped format.}

\item{signature_version}{`string` Signature version to use.}

\item{dataset_selection}{`logical` Whether to select specific signature DBs.}

\item{proteomics_datasets}{`string` Ids for proteomics dataset based signature
DB section.}

\item{perturbation_datasets}{`string` Ids for perturbation dataset based signature
DB section.}
}
\value{
`data.table` ReMEA scores and p values for selected signature DBs.
}
\description{
Same as main version except dofutures is not implemented.
Performs signture delta ranks and delta zscores iteratively across
selected signature DBs. Determined by pertubation type, tumour type and siganture
set.
}
\examples{
response_marker_enrichment_analysis(protein_data,
marker_type=c("DRUG", "RNAi", "CRISPR"),
tumour_type=c("haem","solid"))
}
