% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature_selector.R
\name{select_signature}
\alias{select_signature}
\title{Select signature}
\usage{
select_signature(signature_version)
}
\arguments{
\item{signature_version}{`string` ID of signature list version}
}
\value{
ReMEA signature list.
}
\description{
Select signature
}
