% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_response_marker_enrichment_analysis.R
\name{signature_enrichment}
\alias{signature_enrichment}
\title{Signature enrichment}
\usage{
signature_enrichment(
  protein_data,
  signatures,
  signature.type = "signature_id",
  protein_id_col = NULL,
  analysis_col = NULL
)
}
\arguments{
\item{protein_data}{A data.table containing proteomics differential-testing
results. A data.frame is also accepted and will be converted to a
data.table. Proteins must be annotated with their UniProt Entry Names.
The protein identifier column can be specified using protein_id_col; if
omitted, the first column is used. The numeric column to analyse, such as
fold change, can be specified using analysis_col; if omitted, the second
column is used.}

\item{signatures}{`data.table` Signature DB with response protein signatures.
This can be found in the ReMEA NAMESPACE. e.g. ReMEA::list_of_signature_lists}

\item{signature.type}{`string` Naming variable for signatures. Main purpose if for
response_marker_enrichment_analysis, where this indicates signature DB ids.}

\item{protein_id_col}{Column that indicates protein IDs. For package signatures, this
must be UniProt entry name. If NULL, will default to first column.}

\item{analysis_col}{Data column that is to be analysed. If NULL, will default to second column.}
}
\value{
`data.table` with enrichment results
}
\description{
Main marker enrichment function for ReMEA pipeline.
}
\examples{
signature_enrichemnt(protein_data = protein_data, signatures  = signature_dt, signature.type = "signature_id")
}
