% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remea_analysis.R
\name{complete_ReMEA_analysis}
\alias{complete_ReMEA_analysis}
\title{Complete ReMEA analysis of all perturbation types}
\usage{
complete_ReMEA_analysis(
  protein_data,
  tumour_type,
  signature_version = "CellLines",
  save_xlxs = FALSE
)
}
\arguments{
\item{protein_data}{`data.frame` Proteomics differential testing results. Protein name
must be included as protein_acc. Fold change data must be in column 2.}

\item{tumour_type}{`string` Whether to use markers from correlation analysis of solid tumours
or non-solid tumours only. If multiple passed, these will be analysed in looped format.}

\item{signature_version}{`string` Signature version to use.}

\item{save_xlxs}{`logical` Whether or not to save scores to xlxs file.}
}
\value{
ggplot object
}
\description{
Performs full perturbation analysis of given proteomics data and selected tumour type.
}
\details{
Returns a `list` with 8 slots. Includes ReMEA analysis for all three perturbation
types and averaged analysis of RNAi and CRISPR. Both combined scores and individual
db results are returned.
}
